requirements.txt
Flask==1.0.2
Flask-Rauth
Flask-Cors==3.0.7
gunicorn==19.9.0
requests
